+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 * Introduce limited alpha support for Ambient Weather devices (only WS1965)
 * Implement support for WN1900 / WN1910 / WN1920 Ecowitt consoles
-* Updates for easier first time setup for MAC and path settings
 * Add support for WN67 5-in-1 weather sensor array
+* Resolve a bug where the comparison PASSKEY was generated incorrectly
+* Updates in Plugin UI and log messages for easier first-time setup
 * Add more documentation on setting name overrides for a sensor
 
 ### 2.7.0-beta